Reorganize internal structure
Move the main logic for the keyboard-select-picker into a mixin to help
abstract things and make the move away from ember-keyboard-shortcuts
dependency.
