+++ app/components/list-picker.js
@@ -3,7 +3,7 @@ import SelectPickerMixin from 'ember-cli-select-picker/mixins/select-picker';
 
 var I18nProps = (Ember.I18n && Ember.I18n.TranslateableProperties) || {};
 
-var ListPickerComponent = Ember.Component.extend(
+export default Ember.Component.extend(
   SelectPickerMixin, I18nProps,
   {
     classNames: ['select-picker', 'list-picker'],
@@ -12,5 +12,3 @@ var ListPickerComponent = Ember.Component.extend(
     nativeMobile: false
   }
 );
-
-export default ListPickerComponent;
