+++ app/components/select-picker.js
@@ -3,7 +3,7 @@ import SelectPickerMixin from 'ember-cli-select-picker/mixins/select-picker';
 
 var I18nProps = (Ember.I18n && Ember.I18n.TranslateableProperties) || {};
 
-var SelectPickerComponent = Ember.Component.extend(
+export default Ember.Component.extend(
   SelectPickerMixin, I18nProps, {
 
   nothingSelectedMessage: 'Nothing Selected',
@@ -57,5 +57,3 @@ var SelectPickerComponent = Ember.Component.extend(
     }
   }
 });
-
-export default SelectPickerComponent;
